rb/aa - moved code from app.js to map.js
