+++ src/App.js
@@ -13,21 +13,6 @@ import {
   InfoWindow
 } from "react-google-maps";
 
-function MyMap() {
-  
-
-  return (
-    <GoogleMap
-      defaultZoom={10}
-      defaultCenter={{ lat: 34.4134, lng: -119.8433 }}
-      //defaultOptions={{ styles: mapStyles }}
-    >
-      
-    </GoogleMap>
-  );
-}
-
-const MapWrapped = withScriptjs(withGoogleMap(MyMap));
 
 export default function App() {
   const [isSignedIn, setIsSignedIn] = useState(null);
@@ -77,16 +62,7 @@ export default function App() {
         <PrivateRoute exact path="/profile" Component={Profile} />
         <Route path="/" render={PageNotFound} />
       </Switch>
-      <div style={{ width: "50vw", height: "50vh" }}>
-      <MapWrapped
-        googleMapURL={`https://maps.googleapis.com/maps/api/js?v=3.exp&libraries=geometry,drawing,places&key=${
-          process.env.REACT_APP_GOOGLE_KEY
-        }`}
-        loadingElement={<div style={{ height: `100%` }} />}
-        containerElement={<div style={{ height: `100%` }} />}
-        mapElement={<div style={{ height: `100%` }} />}
-      />
-    </div>
+     
     </BrowserRouter>
   );
 }
